Relocate patch to lib directory
